Clarify instructions; replace secondSmallest
**Changes in this PR**
1. Replace `secondSmallest` with a more straightforward problem: find the smallest number. We have used `secondSmallest` to interview staff so it's not appropriate for Module 1.
2. Add examples to all instructions
3. Rephrase prompts that seem to suggest implementations that will give people trouble (ex: removing elements from an array in place)
4. Remove test cases and instructions for `isOdd` to do with handling non-number input.

Regarding 4), I'd perfer to leave out this edge case because it can be a rabbit hole that we are not fully testing anyway.
```
isOdd(NaN); // false
isOdd("hello"); // false
isOdd("3"); // should be false by our instructions, but will return true for most implementations
```
It's the last case (a string that can be coerced into a number) that I think will pull fellows down into a rabbit hole.
The implementation that most people will use will handle the other test cases, so I recommend we remove the edge case tests.